# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nazrein API
  version: 1.0.0
  description: REST API for Nazrein video title and thumbnail changelog tracker.
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.nazrein.dev
    description: Production
paths:
...
  /api/v1/request:
    get:
      summary: Get video requests made by the current user
      operationId: getUserVideoRequests
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoRequestsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Submit a video tracking request
      operationId: createVideoRequest
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoRequestInput"
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "201":
          $ref: "#/components/responses/MessageOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/request/{id}:
    delete:
      summary: Delete a video request by ID
      operationId: deleteVideoRequest
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/bookmark/{id}:
    post:
      summary: Bookmark a video
      operationId: createBookmark
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "201":
          $ref: "#/components/responses/MessageOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Remove a bookmark
      operationId: deleteBookmark
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/MessageOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nazrein_session
      description: User session cookie, set by the Google OAuth callback
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: nazrein_admin_session
      description: Admin session cookie, set by the admin Google OAuth callback
  responses:
    MessageOK:
      description: Action completed successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    BadRequest:
      description: Invalid query, path, or body parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Action forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    VideoRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED]
        link:
          type: string
        youtube_id:
          type: string
        user_id:
          type: string
        processed_by:
          type: string
          format: uuid
          nullable: true
        processed_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - status
        - link
        - youtube_id
        - user_id
        - created_at
        - updated_at
    VideoRequestsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VideoRequest"
      required:
        - data
...
    CreateVideoRequestInput:
      type: object
      properties:
        link:
          type: string
      required:
        - link
...